Split games & tournament control panels
