+++ Rheda/config/routes.php
@@ -37,14 +37,17 @@ return [ // Omit trailing slashes in keys when possible
     '/sortition'      => 'Sortition',
     '/sortition/(?<seed>[0-9a-f]+)' => 'Sortition',
 
-    '/tourn'                                               => 'StartTournament',
-    '/tourn/(?<action>dropLastRound)/(?<hash>[0-9a-f]+)'   => 'StartTournament',
-    '/tourn/(?<action>shuffledSeating)'                    => 'StartTournament',
-    '/tourn/(?<action>manualSeating)'                      => 'StartTournament',
-    '/tourn/(?<action>swissSeating)'                       => 'StartTournament',
-    '/tourn/(?<action>startTimer)'                         => 'StartTournament',
-    '/tourn/(?<action>toggleHideResults)'                  => 'StartTournament',
-    '/tourn/(?<action>finalizeSessions)'                   => 'StartTournament',
+    '/tourn'                                               => 'TournamentControlPanel',
+    '/tourn/(?<action>dropLastRound)/(?<hash>[0-9a-f]+)'   => 'TournamentControlPanel',
+    '/tourn/(?<action>shuffledSeating)'                    => 'TournamentControlPanel',
+    '/tourn/(?<action>manualSeating)'                      => 'TournamentControlPanel',
+    '/tourn/(?<action>swissSeating)'                       => 'TournamentControlPanel',
+    '/tourn/(?<action>startTimer)'                         => 'TournamentControlPanel',
+    '/tourn/(?<action>toggleHideResults)'                  => 'TournamentControlPanel',
+    '/tourn/(?<action>finalizeSessions)'                   => 'TournamentControlPanel',
+
+    '/games'                                               => 'GamesControlPanel',
+    '/games/(?<action>dropLastRound)/(?<hash>[0-9a-f]+)'   => 'GamesControlPanel',
 
     '/penalties'       => 'Penalties',
     '/penalties/(?<action>apply)' => 'Penalties',
